requirements.txt
colorama
pyautogui
pillow
opencv-python
imagehash